+++ cli/src/commands/auth/utils.ts
@@ -1,7 +1,4 @@
-import { existsSync, readFileSync } from 'node:fs';
-import yaml from 'js-yaml';
-import pc from 'picocolors';
-import { config, configFile } from '../../core/config.js';
+import { config } from '../../core/config.js';
 
 interface DeviceAuthResponse {
   deviceCode: string;
@@ -107,13 +104,4 @@ export const startPollingForAccessToken = async ({
   }
 };
 
-export const checkConfigFile = () => {
-  if (existsSync(configFile)) {
-    const data = yaml.load(readFileSync(configFile, 'utf8'));
-    const loginData = JSON.parse(JSON.stringify(data));
-    if (loginData && loginData?.expiresAt && new Date(loginData.expiresAt) > new Date()) {
-      console.log(pc.green('You are already logged in.'));
-      process.exit(0);
-    }
-  }
-};
+
